Changes:  - Introduced `SimulationManager` class that substitutes the role of the `run_febid()` function. It solely manages thread orchestration, simulation initialization, stopping and finish.  - `SimulationContext` class was stripped off Statistics, StructureSaver, SynchronizationHelper and the printingThread property as they all migrated to the `SimulationManager`.  - The main simulation event loop and cell-filling event functions were moved to the `engine.py`.
