+++ febid/ui/ui_shell.py
@@ -525,7 +525,7 @@ class MainPanel(QMainWindow, UI_MainPanel):
         self.stop_febid_button.setVisible(True)
 
         def wait_for_success():
-            flag = self.session_handler.starter.context.syncHelper
+            flag = self.session_handler.starter.syncHelper
             flag.event.wait()
             if flag.is_success:
                 self.on_finish('Simulation finished')
@@ -565,7 +565,7 @@ class MainPanel(QMainWindow, UI_MainPanel):
             if self.viz.isVisible():
                 return
         self.session_handler.starter.context.process.displayed_data = self.displayed_data # enables data acquisition from GPU
-        self.viz = RenderWindow(self.session_handler.starter.context.process, self.session_handler.starter.context.syncHelper,
+        self.viz = RenderWindow(self.session_handler.starter.context.process, self.session_handler.starter.syncHelper,
                                 displayed_data=self.displayed_data, show=True, app=self.app)
         self.viz.start(frame_rate=self.frame_rate)
 
@@ -703,7 +703,6 @@ class MainPanel(QMainWindow, UI_MainPanel):
         self.start_febid_button.setVisible(True)
         self.stop_febid_button.setVisible(False)
         self.groupBox_visualization.setVisible(False)
-        self.session_handler.starter.context.syncHelper.reset()
         self.statusBar().showMessage(message)
 
     def on_close(self):
